+++ src/core/DynamicPathTracingSceneGenerator.js
@@ -1,6 +1,6 @@
 import { BufferGeometry, MeshBasicMaterial, BufferAttribute, Mesh } from 'three';
 import { MeshBVH, SAH } from 'three-mesh-bvh';
-import { StaticGeometryGenerator } from './StaticGeometryGenerator.js';
+import { StaticGeometryGenerator } from './utils/StaticGeometryGenerator.js';
 import { setCommonAttributes, getGroupMaterialIndicesAttribute } from '../utils/GeometryPreparationUtils.js';
 
 const dummyMaterial = new MeshBasicMaterial();
