+++ src/core/utils/convertToStaticGeometry.js
@@ -1,4 +1,5 @@
-import { BufferAttribute, BufferGeometry, Matrix3, Matrix4, Vector3, Vector4 } from 'three';
+import { BufferGeometry, Matrix3, Matrix4, Vector3, Vector4 } from 'three';
+import { copyAttributeContents, createAttributeClone, validateAttributes } from './BufferAttributeUtils.js';
 
 const _positionVector = /*@__PURE__*/ new Vector3();
 const _normalVector = /*@__PURE__*/ new Vector3();
@@ -12,28 +13,6 @@ const _skinIndex = /*@__PURE__*/ new Vector4();
 const _skinWeight = /*@__PURE__*/ new Vector4();
 const _matrix = /*@__PURE__*/ new Matrix4();
 const _boneMatrix = /*@__PURE__*/ new Matrix4();
-
-// Confirms that the two provided attributes are compatible
-function validateAttributes( attr1, attr2 ) {
-
-	if ( ! attr1 && ! attr2 ) {
-
-		return;
-
-	}
-
-	const sameCount = attr1.count === attr2.count;
-	const sameNormalized = attr1.normalized === attr2.normalized;
-	const sameType = attr1.array.constructor === attr2.array.constructor;
-	const sameItemSize = attr1.itemSize === attr2.itemSize;
-
-	if ( ! sameCount || ! sameNormalized || ! sameType || ! sameItemSize ) {
-
-		throw new Error();
-
-	}
-
-}
 
 // A version of "SkinnedMesh.boneTransform" for normals
 function boneNormalTransform( mesh, index, target ) {
@@ -151,47 +130,6 @@ function invertGeometry( geometry ) {
 	}
 
 	return geometry;
-
-}
-
-// target offset is the number of elements in the target buffer stride to skip before copying the
-// attributes contents in to.
-export function copyAttributeContents( attr, target, targetOffset = 0 ) {
-
-	if ( attr.isInterleavedBufferAttribute ) {
-
-		const itemSize = attr.itemSize;
-		for ( let i = 0, l = attr.count; i < l; i ++ ) {
-
-			const io = i + targetOffset;
-			target.setX( io, attr.getX( i ) );
-			if ( itemSize >= 2 ) target.setY( io, attr.getY( i ) );
-			if ( itemSize >= 3 ) target.setZ( io, attr.getZ( i ) );
-			if ( itemSize >= 4 ) target.setW( io, attr.getW( i ) );
-
-		}
-
-	} else {
-
-		const array = target.array;
-		const cons = array.constructor;
-		const byteOffset = array.BYTES_PER_ELEMENT * attr.itemSize * targetOffset;
-		const temp = new cons( array.buffer, byteOffset, attr.array.length );
-		temp.set( attr.array );
-
-	}
-
-}
-
-// Clones the given attribute with a new compatible buffer attribute but no data
-export function createAttributeClone( attr, countOverride = null ) {
-
-	const cons = attr.array.constructor;
-	const normalized = attr.normalized;
-	const itemSize = attr.itemSize;
-	const count = countOverride === null ? attr.count : countOverride;
-
-	return new BufferAttribute( new cons( itemSize * count ), itemSize, normalized );
 
 }
