+++ src/core/utils/mergeGeometries.js
@@ -1,0 +1,161 @@
+import { BufferAttribute, BufferGeometry } from 'three';
+import { copyAttributeContents, createAttributeClone } from './BufferAttributeUtils.js';
+
+// Modified version of BufferGeometryUtils.mergeBufferGeometries that ignores morph targets and updates a attributes in place
+export function mergeGeometries( geometries, options = { useGroups: false, updateIndex: false, skipAttributes: [] }, targetGeometry = new BufferGeometry() ) {
+
+	const isIndexed = geometries[ 0 ].index !== null;
+	const { useGroups = false, updateIndex = false, skipAttributes = [] } = options;
+
+	const attributesUsed = new Set( Object.keys( geometries[ 0 ].attributes ) );
+	const attributes = {};
+
+	let offset = 0;
+
+	targetGeometry.clearGroups();
+	for ( let i = 0; i < geometries.length; ++ i ) {
+
+		const geometry = geometries[ i ];
+		let attributesCount = 0;
+
+		// ensure that all geometries are indexed, or none
+		if ( isIndexed !== ( geometry.index !== null ) ) {
+
+			throw new Error( 'StaticGeometryGenerator: All geometries must have compatible attributes; make sure index attribute exists among all geometries, or in none of them.' );
+
+		}
+
+		// gather attributes, exit early if they're different
+		for ( const name in geometry.attributes ) {
+
+			if ( ! attributesUsed.has( name ) ) {
+
+				throw new Error( 'StaticGeometryGenerator: All geometries must have compatible attributes; make sure "' + name + '" attribute exists among all geometries, or in none of them.' );
+
+			}
+
+			if ( attributes[ name ] === undefined ) {
+
+				attributes[ name ] = [];
+
+			}
+
+			attributes[ name ].push( geometry.attributes[ name ] );
+			attributesCount ++;
+
+		}
+
+		// ensure geometries have the same number of attributes
+		if ( attributesCount !== attributesUsed.size ) {
+
+			throw new Error( 'StaticGeometryGenerator: Make sure all geometries have the same number of attributes.' );
+
+		}
+
+		if ( useGroups ) {
+
+			let count;
+			if ( isIndexed ) {
+
+				count = geometry.index.count;
+
+			} else if ( geometry.attributes.position !== undefined ) {
+
+				count = geometry.attributes.position.count;
+
+			} else {
+
+				throw new Error( 'StaticGeometryGenerator: The geometry must have either an index or a position attribute' );
+
+			}
+
+			targetGeometry.addGroup( offset, count, i );
+			offset += count;
+
+		}
+
+	}
+
+	// merge indices
+	if ( isIndexed ) {
+
+		let forceUpdateIndex = false;
+		if ( ! targetGeometry.index ) {
+
+			let indexCount = 0;
+			for ( let i = 0; i < geometries.length; ++ i ) {
+
+				indexCount += geometries[ i ].index.count;
+
+			}
+
+			targetGeometry.setIndex( new BufferAttribute( new Uint32Array( indexCount ), 1, false ) );
+			forceUpdateIndex = true;
+
+		}
+
+		if ( updateIndex || forceUpdateIndex ) {
+
+			const targetIndex = targetGeometry.index;
+			let targetOffset = 0;
+			let indexOffset = 0;
+			for ( let i = 0; i < geometries.length; ++ i ) {
+
+				const geometry = geometries[ i ];
+				const index = geometry.index;
+				if ( skipAttributes[ i ] !== true ) {
+
+					for ( let j = 0; j < index.count; ++ j ) {
+
+						targetIndex.setX( targetOffset, index.getX( j ) + indexOffset );
+						targetOffset ++;
+
+					}
+
+				}
+
+				indexOffset += geometry.attributes.position.count;
+
+			}
+
+		}
+
+	}
+
+	// merge attributes
+	for ( const name in attributes ) {
+
+		const attrList = attributes[ name ];
+		if ( ! ( name in targetGeometry.attributes ) ) {
+
+			let count = 0;
+			for ( const key in attrList ) {
+
+				count += attrList[ key ].count;
+
+			}
+
+			targetGeometry.setAttribute( name, createAttributeClone( attributes[ name ][ 0 ], count ) );
+
+		}
+
+		const targetAttribute = targetGeometry.attributes[ name ];
+		let offset = 0;
+		for ( let i = 0, l = attrList.length; i < l; i ++ ) {
+
+			const attr = attrList[ i ];
+			if ( skipAttributes[ i ] !== true ) {
+
+				copyAttributeContents( attr, targetAttribute, offset );
+
+			}
+
+			offset += attr.count;
+
+		}
+
+	}
+
+	return targetGeometry;
+
+}
